fix: исправить замечания
